chore: 모듈 적용
