+++ public/index.html
@@ -5,6 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <script type="module" src="src/main.js" defer></script>
+    <script type="module" src="src/dispatchMessage.js" defer></script>
     <link rel="stylesheet" href="style.css">
     <title>마션</title>
     <h1>마션</h1>
